agent/pyproject.toml
[project]
name = "agent-utilities"
version = "0.1.0"
description = "Algolia Agent Studio CLI for TCG Search"
authors = [
    {name = "Chuck Meyer",email = "chuck.meyer@algolia.com"}
]
requires-python = ">=3.14"
dependencies = [
    "algolia-agent @ git+https://github.com/algolia-samples/algolia-agent-cli.git",
    "python-dotenv (>=1.0.0,<2.0.0)",
]


[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
